live: fixed crash when secret event was deleted
Essentially, we were scrying for our remote scry revision number for a
secret event, which do not have them. The scry was crashing which
wouldn't allow the user to press "yes" when asked to delete.

Now, we route around the revision number ask for a secret event.

The outlier case is when a public event changes to secret, but this is
already handled in +change-info.