live: host events populate when they search for themselves
Before the search would display "no events found", but to make it
seamless, the host can now use "find" to search among their own events.

One notable addition includes the $result overwrite in +delete. This is
there to route around a bug where if the host deleted an event and their
result contained it, the "find" button would stop working.