live & live-view: ship entry error handling
previously the backend would crash when given an invalid ship format. now
we've added an error handling apparatus to tell the user when this
happens.

this mainly effected the +compose logic in live.hoon, particularly for
the %find and %invite cases.

no SSEs meant i had to get creative with handling this. i added a
+handle-error arm which now builds the manage and find pages with their
respective error messages. this is unideal, but it works.

broadly, if more of this kind of user feedback needs to be built,
it's probably worth adopting something like Mast.

finally, the backend now prepends a ~ if the user forgot to include it.