live: fixed event search crash
the gall scry failed due to a changed scry path syntax
with the release of `411`. i've updated the syntax, and also refactored
+case to clearly distinguish between a case request and response. previously, both
functions were baked into a single arm, which made updating it
difficult. it was confusing which points to a semantic problem so i
broke it into two arms (+case-request and +case-response) which also
correspond to new respective action pokes.

so now when a ship searches for an event, it sends a local `%find` poke
to their agent, who then sends a `%case-request` to the host ship. the
host receives this and scries the case number for the path in question
and returns it to the requesting ship as a `%case-response`. either the
host ship includes a case with the response, meaning they have discoverable event
at that path so the requesting ship proceeds with performing the remote scry; or there's
no case provided, in which case the requesting ship can assume there are no
events at that path.

i also make a small modification to `live-view` which includes four
random characters instead of three next to the reload button. at some
point in the past this was accidentally modified; four is obviously more
aesthetically correct.