live: sessions support and group/location field additions
a session is a component of an event (event within an event) with a slightly different type signature than a top-level event.

operating on a session works similarly to operating on an event; there's granular update support for specific session fields. sessions live in a map under the info-1:event-1 structure.

you can create/edit/delete a session by passing a [%info <sub-info <sub-session>>] argument.