matcher: %shake function; i.e. matching
Includes the basic matching functions that happen on the host ship.

A guest can either send a positive or negative %shake for some other guest in question.

A positive %shake adds the target guest into the actor's reaches list. If the target already had them in their reaches list, they're both added to each other's matches list, instead.

If the target later sends a positive %shake to the actor, they're cleared from the reaches list and added to matches.

This commit only includes the state updates required to keep track of who's matched with who. The next one will include pokes to the respective guests so they can keep track of their network.