matcher: +shake refactoring and +show completed
I refactored +shake slighly to make it more readable.

Also changed $status to a `(unit ?(%match %incoming %outgoing))` because the binary %reach/%match wasn't enough; a guest needs to know whether they've initiated a match or not. With just %reach, it was unclear on the recipent's ship who initiated it. This was also cause for refactoring +shake.

+shake now passes a %show $dictum to the respective peers so they can update their status to each other.

I also added %poke-ack handling in +agent for both pokes.