matcher: fixed subscription state desynchronization
When someone registered, unregistered, and then re-registered to an event the peers subscription state would not match the host. It was happening because our sss retention policy wasn't updating the rock until every 5th wave. Also, the sss updates were updating peers state by wave, which would sometimes arrive out of sequence causing unnecessary state overwrites.

Now now the host's retention policy only keeps the latest rock, and subscribers branch on whether the subscription is stale or not. If stale, they'll unsubscribe and delete the event for that set of peers; if not, they simply overwrite the set of peers with the latest rock they receive.