live: fixed resubscription bug
the core state modifications weren't being written because all branches of +change-info resulted in an $event and not the cor so i changed it to a nested core to keep the changes.

the %kind branch also wasn't handling the case where the event changes from %secret to %public or %private again, thus subscriptions for %requested and %unregistered guests weren't being reestablished.

finally, I bundled the subscription poke card builders into a single arm: +ask-to-subscribe