live: `$panel` now converts back into a json array
I considered a few different approaches to fixing the problem -
initially starting with converting the frontend session structure, but
this only led to separate instances of what should be the same type.

it also makes more sense to have this be an array, for ease of
modification, and ultimately the backend should have made this structure
a tape in the first place, but this alone does not justify a full state transition.

so we handle this in the middle layer, converting from an array and
back.

also in this commit is a small formatting adjustment for the boilerplate
input handlers.