live: made edit errors more granular
which basically meant pushing the backend errors that would go to the
dojo to the frontend. this required refactoring the +change-info and
+ingest-diff arms pretty significantly.

this commit also includes slight refactoring of where the backend pushes
error messages - before we had +fail and +succ; now there's only +succ
which handles both successes and failures.