live: init an event to over caused crash
now fixed. this was happening because the new core was being written
before +get-our-case scried for existing cases. a case didn't exist
because the event was never published (i.e. setting to over from the
jump just writes it locally). the new core of course had a latch that
was either open or closed so it was passing the filter, causing a crash.

just needed to move the core update below the +get-our-case scry.